Quick implementation of the image flip required for lossless transcode compatibility with VP3. THIS IS AN INCOMPATIBLE BITSTREAM CHANGE. Not really, but alpha 2 theora players and encoders will see the video as upside down.
In theora_encode_YUVin() we invert the image as we copy it to our
internal buffer. However, in theora_decode_YUVout() I've just inverted
the row pointer. If there's a better way to do this, please let me know;
I'd be much happier returning an inverted buffer. Perhaps this can be
done as part of the post-process step, at the expense of always
requiring at least one pp pass?