Various clean-ups and bugfixes.
With this commit, decoder output should be bit-identical to that of the
 reference decoder (for files the reference decoder actually supports).
This time I actually checked.

The "motion vector" problem was actually the golden frame getting blown away by
 the following inter frame in some cases.

There were also some minor problems with the order the loop filtering and
 border padding operations were being applied.
These were resolved by fixing the reference decoder.

Finally, we did not handle an EOB run of length 0 the same way that the
 reference decoder did.
We treated it as a nop (violating some of our design assumptions in the
 process), instead of ending every block in the frame.
This has been corrected.