Snotfish is showing signs of detroying / again, so I'd better commit script work in progress.
These changes to the book generation scripts and configuration files
should prevent future occurrences of a subtle mistake that bit me last
night (training with a mismatched aux partition book that causes one
partition to be unusable; the partition code in the encoder assumes
partition books to be complete and lossless, so that caused all sorts
of hell to break loose).

There's a good chance that this is the book bug that bit Garf when he
was working on low bitrate modes, and it's certainly the root cause of
bitrate management not working properly/at all in AoTuV.  (The bug is
that it's possible to unknowingly create incomplete temporary books
when bootstrap training such that the training results in corrupted
books.  This is a 'process' bug.)