Upgraded to use the work from my libogg2 branch in svn:  http://svn.xiph.org/branches/ogg2-arc
Added setmode() to OggStreamState objects to set discontinuous mode.

Removed temporary include in general.h.

Made a small fix in examples/writ_encoder.py, as it was previously 
attempting to use a packet after it had been passed into the 
OggStreamState object (I believe this is not legal, and made this an 
error in py-ogg2 with an earlier commit).