This solves bug 485 and re-adds ogg_page_checksum_set to ogginternal.h
Tt used to be in include/ogg2/ogg.h but was removed because there are 
no external functions for modifying data, and thus, this is useless to 
the outside.  I don't quite understand how my code has been compiling 
without it all this time, but it's surely needed or others to compile.