A second major upgrade to the math behind the Bessel followers used by the companders.
a) Soft knees are now computed as a freebie during filter following.
No explicit soft-knee means no sqrt() means considerable optimization

b) streamlined the computation pathway of the followers, partially as
an optimization, partially to prepare for adding the scond tier
over/under filters before release

c) found a bug in the behavior of the new peak-follower math (that was
instigated to repair a more serious bug in the original version);
decay was still using a second order bessel which could still ring
about zero.  Restored the peak-hold behavior to near-original
behavior, but now use the same first-order-decay-limited filters as
the RMS follower.

Depending on platform and compander usage, these changes are good for
getting back 10-25% of the CPU.