Merge libtheorabase into libtheoradec. This means you link to just libtheoradec for decoder support and both libtheoraenc and libtheoradec for encoder support.
While the common api is useful on its own, a great deal of the included 
code was not (like the idct implementation). I believe that the 
simplicity of having only two libraries overrides the value of the 
separate, shared base library. Since libtheoradec and libtheorabase
were about the same size, this double the size of the dependency for
people using only the shared API. On the other hand, much of the shared 
API code, like theora_packet_isfoo() and the comment manipulation 
routines, is small enough to poach if the footprint is an issue, or the 
client can just link statically. Likewise, this change adds about 40 kB
to the size of encoder-only applications.

I have updated the autoconf and GNU make ("unix") builds, and 
(hopefully) all the documentation. Someone else will have to
make the corresponding change to the win32 project files.