(from Karl)
Another diff, this time it fixes a memory leak, the memory does get
freed up but only on close or when new headers come in.  So if using
something like ffmpeg2theora as the feed, the memory usage will
accumulate while running.
        
I've also changed an internal structure element name from the previous
patch as that was originally based on a time but really it's based on frame.