Remove the limits on the the video rate target in the encoder example.
I don't know why these are here, but thusnelda doesn't have any rate 
control at all, so they are unnecessary. Furthermore, 2 Mbps is too
low for much HD and larger content, so it's better to remove them
with the expectation that a wider range will be acceptable to the final 
encoder.

The file format itself records the target bitrate in the ident header, 
but this field is limited to 24 bits in units of bits per second, for
a maximum of 16 Mbps. Another reason this field isn't very useful. 

Because of this, the ui will warn about clamping to 16 Mbps if a larger 
value is requested. This is, in fact, what the current code does in
theora_encode_init().