Generalize the structure of the OpenBSD checks a bit.
We add -lcompat to everything, but only if it's needed for ftime(). This 
moves clutter from Makefile.am to the already messy build lines.

We still add -lossaudio iff the target is OpenBSD, but we call it 
OSS_LIBS instead of OBSD_OSSAUDIO which is both shorter and more
generic should other systems need something in the future.