Fix typo in conditional.
Typo in the conditional to check the exit condition in ascii_strcasecmp.
It checks for (!*s1 || !*s1) instead of (!*s1 || !*s2). The typo did
not affect the result of the function as the loop is exited before
changing s1 or s2 anyway.

Fix ported from the copy in the flac library. Thanks to dcb314 for
reporting it here.

Closes https://trac.xiph.org/ticket/2007